+++ packages/bpk-component-button/stories.js
@@ -1,7 +1,14 @@
 import React from 'react';
 import { storiesOf, action } from '@kadira/storybook';
 
+import TOKENS from 'bpk-tokens/tokens/base.common';
+import { withButtonAlignment, withLargeButtonAlignment } from 'bpk-component-icon';
+import SmallLongArrowRightAltIcon from 'bpk-component-icon/sm/long-arrow-right-alt';
+import LargeLongArrowRightAltIcon from 'bpk-component-icon/lg/long-arrow-right-alt';
 import BpkButton from './index';
+
+const AlignedSmallLongArrowRightAltIcon = withButtonAlignment(SmallLongArrowRightAltIcon);
+const AlignedLargeLongArrowRightAltIcon = withLargeButtonAlignment(LargeLongArrowRightAltIcon);
 
 storiesOf('bpk-component-button', module)
   .add('Primary', () => (
@@ -47,6 +54,38 @@ storiesOf('bpk-component-button', module)
       &nbsp;<BpkButton link large onClick={action('large link button clicked')}>Button</BpkButton>
       &nbsp;<BpkButton link large selected onClick={action('large link button selected clicked')}>Selected</BpkButton>
       &nbsp;<BpkButton link large disabled onClick={action('THIS SHOULD NOT HAPPEN')}>Disabled</BpkButton>
+    </div>
+  ))
+  .add('Icon-only button', () => (
+    <div>
+      <div>
+        <BpkButton iconOnly onClick={action('iconOnly button clicked')}>
+          <AlignedSmallLongArrowRightAltIcon fill={TOKENS.colorWhite} />
+        </BpkButton>&nbsp;
+        <BpkButton iconOnly secondary onClick={action('iconOnly button secondary clicked')}>
+          <AlignedSmallLongArrowRightAltIcon fill={TOKENS.buttonSecondaryColor} />
+        </BpkButton>&nbsp;
+        <BpkButton iconOnly selected onClick={action('iconOnly button selected clicked')}>
+          <AlignedSmallLongArrowRightAltIcon fill={TOKENS.colorWhite} />
+        </BpkButton>&nbsp;
+        <BpkButton iconOnly disabled onClick={action('THIS SHOULD NOT HAPPEN')}>
+          <AlignedSmallLongArrowRightAltIcon fill={TOKENS.colorWhite} />
+        </BpkButton>&nbsp;
+      </div>
+      <div>
+        <BpkButton iconOnly large onClick={action('large iconOnly button clicked')}>
+          <AlignedLargeLongArrowRightAltIcon fill={TOKENS.colorWhite} />
+        </BpkButton>&nbsp;
+        <BpkButton iconOnly large secondary onClick={action('large iconOnly button secondary clicked')}>
+          <AlignedLargeLongArrowRightAltIcon fill={TOKENS.buttonSecondaryColor} />
+        </BpkButton>&nbsp;
+        <BpkButton iconOnly large selected onClick={action('large iconOnly button selected clicked')}>
+          <AlignedLargeLongArrowRightAltIcon fill={TOKENS.colorWhite} />
+        </BpkButton>&nbsp;
+        <BpkButton iconOnly large disabled onClick={action('THIS SHOULD NOT HAPPEN')}>
+          <AlignedLargeLongArrowRightAltIcon fill={TOKENS.colorWhite} />
+        </BpkButton>&nbsp;
+      </div>
     </div>
   ))
   .add('Mixture', () => (